file rename: appium-service-configurator.ts => appium-adb-wrapper.ts
