+++ src/electron/platform/android/appium-adb-wrapper.ts
